Add a note about including a private key into *.p12 bundle
This is where I lost 3 hours a few days ago.

There is no place which says that a private key must be a part of *.p12 bundle. I think it's because a person who wrote the current document had this private key in their own keychain and just didn't notice that the private key is needed for e.g. CI.

Actually, there is even no way to create a *.p12 bundle from just the certificates: Keychain export proposes to use only *.p7b format (not password-protected and containing no secrets). The certificates themselves contain no secret information (they're public keys), so it's obvious that some private key is essential.

So, after I included the private key to the *.p12 bundle, the signing process started working like a charm.